% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filehash-SQLite.R
\name{dbDelete,filehashSQLite,character-method}
\alias{dbDelete,filehashSQLite,character-method}
\title{Delete Object}
\usage{
\S4method{dbDelete}{filehashSQLite,character}(db, key, ...)
}
\arguments{
\item{db}{object of class "filehashSQLite"}

\item{key}{character vector of key names}

\item{...}{other arguments (not used)}
}
\description{
Delete an object from the database
}
