% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filehash-SQLite.R
\name{dbDisconnect,filehashSQLite-method}
\alias{dbDisconnect,filehashSQLite-method}
\title{Disconnect from Database}
\usage{
\S4method{dbDisconnect}{filehashSQLite}(conn, ...)
}
\arguments{
\item{conn}{database object}

\item{...}{other arguments (not used)}
}
\description{
Disconnect from Database
}
