% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filehash-SQLite.R
\name{dbExists,filehashSQLite,character-method}
\alias{dbExists,filehashSQLite,character-method}
\title{Check Existence of Key}
\usage{
\S4method{dbExists}{filehashSQLite,character}(db, key, ...)
}
\arguments{
\item{db}{object of class "filehashSQLite"}

\item{key}{character vector of key names}

\item{...}{other arguments (not used)}
}
\description{
Check to see if a key is in the database
}
