% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filehash-SQLite.R
\name{dbFetch,filehashSQLite,character-method}
\alias{dbFetch,filehashSQLite,character-method}
\title{Fetch Object}
\usage{
\S4method{dbFetch}{filehashSQLite,character}(db, key, ...)
}
\arguments{
\item{db}{object of class "filehashSQLite"}

\item{key}{character, key name}

\item{...}{other arguments (not used)}
}
\description{
Retrieve the value associated with a specific key
}
