% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filehash-SQLite.R
\name{dbInsert,filehashSQLite,character-method}
\alias{dbInsert,filehashSQLite,character-method}
\title{Insert Object}
\usage{
\S4method{dbInsert}{filehashSQLite,character}(db, key, value, ...)
}
\arguments{
\item{db}{object of class "filehashSQLite"}

\item{key}{character, key name}

\item{value}{R object}

\item{...}{other arguments (not used)}
}
\description{
Insert a key-value pair into a database
}
