% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filehash-SQLite.R
\name{dbMultiFetch,filehashSQLite,character-method}
\alias{dbMultiFetch,filehashSQLite,character-method}
\title{Fetch Multiple Objects}
\usage{
\S4method{dbMultiFetch}{filehashSQLite,character}(db, key, ...)
}
\arguments{
\item{db}{object of class "filehashSQLite"}

\item{key}{character vector of key names}

\item{...}{other arguments (not used)}
}
\description{
Return (as a named list) the values associated with a vector of keys
}
