% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filehash-SQLite.R
\name{dbUnlink,filehashSQLite-method}
\alias{dbUnlink,filehashSQLite-method}
\title{Unlink Database}
\usage{
\S4method{dbUnlink}{filehashSQLite}(db, ...)
}
\arguments{
\item{db}{object of class "filehashSQLite"}

\item{...}{other arguments (not used)}
}
\description{
Remove a database
}
