% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filehash-SQLite.R
\docType{class}
\name{filehashSQLite}
\alias{filehashSQLite}
\alias{filehashSQLite-class}
\title{Filehash SQLite Class}
\description{
Filehash SQLite Class
}
\section{Slots}{

\describe{
\item{\code{datafile}}{character, full path to the file in which the database should be stored}

\item{\code{dbcon}}{Object of class \dQuote{SQLiteConnection}, a SQLite connection}

\item{\code{drv}}{\sQuote{SQLite} driver}

\item{\code{name}}{character, the name of the database}
}}

\note{
\dQuote{filehashSQLite} databases have a \code{"["} method that can be
used to extract multiple elements in an efficient manner.  The return
value is a list with names equal to the keys passed to \code{"["}.
If there are keys passed to \code{"["} that do not exist in the
database, a warning is given.

The \dQuote{SQLite} format for \code{filehash} uses an ASCII
serialization of the data which could result in some rounding error
for floating point numbers.

Note that if you use keys that are numbers coerced to character
vectors, then you may have trouble with them being coerced to
numeric.  The SQLite database will see these key values and
automatically convert them to numbers.
}
