% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filehash-SQLite.R
\name{[,filehashSQLite,character,ANY,ANY-method}
\alias{[,filehashSQLite,character,ANY,ANY-method}
\title{Fetch Multiple Objects Operator}
\usage{
\S4method{[}{filehashSQLite,character,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{object of class "filehashSQLite"}

\item{i}{index}

\item{j}{index}

\item{...}{other arguments (not used)}

\item{drop}{drop dimensions}
}
\description{
Return (as a named list) the values associated with a vector of keys
}
